pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "pyperclip>=1.9.0",
    "pyqt5>=5.15.11",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.2",
]
